[server/http] move connection setup out of main setup
We need to be able to follow the same connection setup routine outside of the server's http module. Since we can create Config instances (with https://github.com/elastic/kibana/pull/5865) and this portion of initialization does not require the actual kbnServer, we can safely use it to configure other hapi instances so that they will behave similarrly to Kibana.
